# performance: remove unecessary calcs