Results in ~50% speedup in the 2 variable iteration case

Before
```
jeswr@debian:~/n3-benchmark/n3-opt/N3.js$ node perf/N3Store-perf.js 128
N3Store performance test
- Adding 2097152 triples to the default graph: 1.239s
* Memory usage for triples: 147MB
- Finding all 2097152 triples in the default graph 16384 times (0 variables): 4.473s
- Finding all 2097152 triples in the default graph 32768 times (1 variable): 1.357s
- Finding all 2097152 triples in the default graph 49152 times (2 variables): 1.453s

- Adding 1048576 quads: 782.903ms
* Memory usage for quads: 133MB
- Finding all 1048576 quads 131072 times: 1.223s
```

After
```
jeswr@debian:~/n3-benchmark/n3-opt/N3.js$ node perf/N3Store-perf.js 128
N3Store performance test
- Adding 2097152 triples to the default graph: 1.184s
* Memory usage for triples: 145MB
- Finding all 2097152 triples in the default graph 16384 times (0 variables): 4.571s
- Finding all 2097152 triples in the default graph 32768 times (1 variable): 1.046s
- Finding all 2097152 triples in the default graph 49152 times (2 variables): 960.889ms

- Adding 1048576 quads: 803.555ms
* Memory usage for quads: 123MB
- Finding all 1048576 quads 131072 times: 816.152ms
```